* misc: added a project file for Microsoft VisualStudio to be able to build   FlexCat for Windows systems. A custom implementation of (v)asprintf had to   be added as well, since VS is too dump to know these functions.

+++ src/scanct.c
@@ -83,7 +83,7 @@ int ScanCTFile(char *ctfile)
             case ';':
                 if(CopyNEWs == TRUE)
                 {
-                    if(cs && Strnicmp(line, Old_Msg_New, strlen(Old_Msg_New)) == 0)
+                    if(cs && Strnicmp(line, Old_Msg_New, (int)strlen(Old_Msg_New)) == 0)
                     {
                         cs->NotInCT = TRUE;
                     }
